In progress of testing PS4. Open loop locomotion
Current issue:
1. Conversion to velocity from PS4 is not correct, especially when both sticks are at their extremes.
2. Left stick alone at max only reaches to 70% duty cycle.

+++ main/Motor.ino
@@ -172,7 +172,7 @@ void forward_hard_coded(double initialPWM, double maxPWM, double rampTimeMs, dou
 // To clamp motor input calculated from PS4 analog stick. Clamping is through a ramping function.
 void input_shaping(double (&wheelMotorInputs) [4], double (&previousWheelMotorInputs) [4], MotorWithEncoder& UL_Motor) {
     // Assume all maxPwmIncrement (defined in class) is the same across all wheel motors
-    static int maxPwmIncrement = UL_Motor.maxPwmIncrement;
+    static double maxPwmIncrement = UL_Motor.maxPwmIncrement;
 
     // Using for loop, go through each motor inputs and apply ramping function
     for (int i = 0; i < 4; i++) {
